Train a MLP on a small subset of CellXGene
In preparation for a larger training run, we built a small training and test set
from the well represented classes in CellXGene (377 out of >700).  This is an initial
training run to get an end-to-end classifier working. The goal is to

1. understand the nuances of the entire dataset
2. create an initial training set
3. create an initial benchmark model
4. understand how to scale our model to a larger subset

Results are in Weights and Biases: https://wandb.ai/honicky/cellxgene-mlp-v1?nw=nwuserhonicky